+++ src/aks-preview/azext_aks_preview/_client_factory.py
@@ -39,6 +39,10 @@ def cf_machines(cli_ctx, *_):
     return get_container_service_client(cli_ctx).machines
 
 
+def cf_operations(cli_ctx, *_):
+    return get_container_service_client(cli_ctx).operation_status_result
+
+
 def cf_maintenance_configurations(cli_ctx, *_):
     return get_container_service_client(cli_ctx).maintenance_configurations