+++ src/aks-preview/azext_aks_preview/_params.py
@@ -931,6 +931,8 @@ def load_arguments(self, _):
             is_preview=True,
             arg_type=get_enum_type(health_probe_modes),
         )
+        # virtual machines
+        c.argument("vm_sizes", is_preview=True)
 
     with self.argument_context("aks update") as c:
         # managed cluster paramerters
@@ -1540,6 +1542,8 @@ def load_arguments(self, _):
             validator=validate_gateway_prefix_size,
             is_preview=True,
         )
+        # virtual machines
+        c.argument("vm_sizes", is_preview=True)
 
     with self.argument_context("aks nodepool update") as c:
         c.argument(
@@ -1649,6 +1653,16 @@ def load_arguments(self, _):
             required=True,
             help="Space-separated machine names to delete.",
         )
+
+    with self.argument_context("aks nodepool manual-scale add") as c:
+        c.argument("vm_sizes", is_preview=True)
+
+    with self.argument_context("aks nodepool manual-scale update") as c:
+        c.argument("current_vm_sizes", is_preview=True)
+        c.argument("vm_sizes", is_preview=True)
+
+    with self.argument_context("aks nodepool manual-scale delete") as c:
+        c.argument("current_vm_sizes", is_preview=True)
 
     with self.argument_context("aks machine") as c:
         c.argument("cluster_name", help="The cluster name.")